Use the same container for all fragments
