+++ example/android/app/src/main/res/layout/activity_main.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.airbnb.android.react.navigation.FragmentContainerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.airbnb.android.react.navigation.ScreenCoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/content"
     android:layout_width="match_parent"
     android:layout_height="match_parent"/>
