+++ lib/android/src/main/java/com/airbnb/android/react/navigation/ReactActivity.java
@@ -4,7 +4,6 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.Gravity;
 import android.view.ViewGroup;
-import android.widget.FrameLayout;
 
 public abstract class ReactActivity extends ReactAwareActivity implements ScreenCoordinatorComponent {
   private static final String TAG = ReactActivity.class.getSimpleName();
@@ -14,7 +13,7 @@ public abstract class ReactActivity extends ReactAwareActivity implements Screen
   @Override
   protected void onCreate(@Nullable Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    FrameLayout container = new FrameLayout(this);
+    ScreenCoordinatorLayout container = new ScreenCoordinatorLayout(this);
     container.setLayoutParams(
         new ViewGroup.LayoutParams(
             ViewGroup.LayoutParams.MATCH_PARENT,
