+++ lib/android/src/main/res/layout/activity_tab_2.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <FrameLayout
+    <com.airbnb.android.react.navigation.ScreenCoordinatorLayout
         android:id="@+id/content"
         android:layout_width="match_parent"
         android:layout_height="0dp"
